Replace Spaces and Add Backup File
I added a simple loop within the readin loop to try to replace the underscores with spaces, but it's not printing out anymore? Anyway I'll fix it later or worst case scenario we can restore an earlier version of the repo. Also I added a backup file for me to play around with without me fucking up Katie's files, lol.